HPCC-13275 Split the child dependency spotting from the resourcer
This change is part of the preparation for improving the execution of conditional
expressions.  By splitting the child dataset spotting from the resourcing it
allows other transformations to also be applied within graph "resourcing".

The change looks substantial, but some of that is moving code and refactoring it.
The main changes are:
- Split the child dataset spotting into a separate stage
- Move some logic into the the options class which is shared between stages
- Move the dependency tracking code into a separate class, and reduce the lifetime
  of some of the variables.
- Allow graph results to be reused by the resourcer
- Introduce the concept of a _graphLocal_ action which is only executed when required
  by another activity.

Signed-off-by: Gavin Halliday <gavin.halliday@lexisnexis.com>